format for tweet; update comments
update the formatting of tweets to take into account the various types of warm-up handling we now do. we also no longer include cooldown time information, because it's just not that interesting.

the duration format we use in tweets is now like "1h 20m", which is friendlier to read than "1:20" yet doesn't take up much space

add a lot of comments to explain the different types of warmup time handling

throw exception on failing to take photos too many times. it will result in going on to the next round.