process config file at the end of every round
previously we were only reading the config file once, at the start of the program. it would be much more useful to read it after every round so the tuning values could be adjusted without having to restart the program.
still only read the status file once at the start of the program. we don't need to support reading it in the middle